Refactored Spotify related classes into their own files
