+++ spoti-friends/src/spotify/models/SpotifyResource.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+/// The `SpotifyResource` protocol ensures each abiding object has a well-defined spotifyUri attribute.
+protocol SpotifyResource {
+    var spotifyUri: String { get }
+    var name: String { get }
+}
+
+/// This extensions defines the function that returns the `spotifyUri` for all `spotifyResource` objects.
+extension SpotifyResource {
+    /// Returns the `spotifyUri` for this `spotifyResource`.
+    func getSpotifyUri() -> String {
+        return self.spotifyUri  // might need to omit the 'self' if it binds itself to the `SpotifyProfile` object
+    }
+}
